users/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from users.models import CustomUser


class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = "__all__"  
