requirements.txt
asgiref==3.3.4
beautifulsoup4==4.10.0
boto3==1.17.101
botocore==1.20.101
cachetools==4.2.4
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cssselect==1.1.0
cssutils==2.3.0
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap5==21.3
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-light==0.1.0.post3
django-queryset-csv==1.1.0
django-seed==0.3.1
django-storages==1.11.1
djangorestframework==3.12.4
Faker==9.6.0
fontawesome==5.10.1.post1
fpdf==1.7.2
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
jmespath==0.10.0
lxml==4.6.4
markdown2==2.4.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.29
numpy==1.21.4
Pillow==8.2.0
premailer==3.10.0
protobuf==4.21.1
psycopg2==2.9.1
Pygments==2.9.0
PyPDF2==1.26.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
reportlab==3.6.3
requests==2.25.1
requests-toolbelt==0.9.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
text-unidecode==1.3
toposort==1.7
unicodecsv==0.14.1
urllib3==1.26.5
whitenoise==5.2.0
